+++ tests/locales/en_US/genderTest.php
@@ -1,0 +1,31 @@
+<?php
+
+$l = [
+    "profession" => [
+        "male" => "He is a developer",
+        "female" => "She is a developer",
+        "neutral" => "They are a developer"
+    ],
+    "purchase" => [
+        "male" => [
+            1 => "He bought one item",
+            2 => "He bought two items",
+            5 => "He bought many items"
+        ],
+        "female" => [
+            1 => "She bought one item",
+            2 => "She bought two items",
+            5 => "She bought many items"
+        ],
+        "neutral" => [
+            1 => "They bought one item",
+            2 => "They bought two items",
+            5 => "They bought many items"
+        ]
+    ],
+    "welcome" => [
+        "male" => "Welcome Mr. {name}",
+        "female" => "Welcome Mrs. {name}",
+        "neutral" => "Welcome {name}"
+    ]
+];